challange1/api/requirements.txt
fastapi
pydantic
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-storage
protobuf
uvicorn[standard]